Enforce conversation retention minimum in resource
